+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## CHAGNELOG
 
+### 3.1.0
+
+* rename `faq_best_reply` to `faqBestReplyThreshold`
+* rename `faq_sugg_reply` to `faqSuggReplyThreshold`
+
 ## 2.1.0
 * enable `faq_best_reply`, `faq_sugg_reply` in `faq` and `conversation` interfaces.